Added Endpoint authorization

+++ src/main/java/SelfBankingSystem/SelfBankingSystem/account/AccountController.java
@@ -1,6 +1,9 @@
 package SelfBankingSystem.SelfBankingSystem.account;
 
+import SelfBankingSystem.SelfBankingSystem.customer.Customer;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -22,12 +25,22 @@ public class AccountController {
 
     @GetMapping(path = "{customerId}")
     public List<Account> getAllAccountsByCustomerId(@PathVariable(value = "customerId") Long customerId){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Customer target = (Customer) authentication.getPrincipal();
+        if(target.getId() != customerId){
+            throw new IllegalStateException("Invalid Request!");
+        }
         return accountService.getAllAccountsByCustomerId(customerId);
     }
 
     @PostMapping(path="{customerId}")
     public void createAccount(@PathVariable(value="customerId") Long customerId,
                               @RequestBody Account account){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Customer target = (Customer) authentication.getPrincipal();
+        if(target.getId() != customerId){
+            throw new IllegalStateException("Invalid Request!");
+        }
         accountService.createAccount(customerId, account);
     }
 
@@ -35,6 +48,11 @@ public class AccountController {
     public void depositAccount(@PathVariable(value="customerId") Long customerId,
                                @RequestParam(required = true) Integer accountType,
                                @RequestParam(required = true) Integer amount){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Customer target = (Customer) authentication.getPrincipal();
+        if(target.getId() != customerId){
+            throw new IllegalStateException("Invalid Request!");
+        }
         accountService.depositAccount(customerId, accountType, amount);
     }
 
@@ -42,18 +60,33 @@ public class AccountController {
     public void withdrawAccount(@PathVariable(value="customerId") Long customerId,
                                @RequestParam(required = true) Integer accountType,
                                @RequestParam(required = true) Integer amount){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Customer target = (Customer) authentication.getPrincipal();
+        if(target.getId() != customerId){
+            throw new IllegalStateException("Invalid Request!");
+        }
         accountService.withdrawAccount(customerId, accountType, amount);
     }
 
     @PutMapping(path = "{customerId}/active")
     public void lockOrUnlockAccount(@PathVariable(value="customerId") Long customerId,
                                 @RequestParam(required = true) Integer accountType){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Customer target = (Customer) authentication.getPrincipal();
+        if(target.getId() != customerId){
+            throw new IllegalStateException("Invalid Request!");
+        }
         accountService.lockOrUnlockAccount(customerId, accountType);
     }
 
     @DeleteMapping(path = "{customerId}")
     public void deleteAccount(@PathVariable(value="customerId") Long customerId,
                                 @RequestParam(required=true) Integer accountType){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Customer target = (Customer) authentication.getPrincipal();
+        if(target.getId() != customerId){
+            throw new IllegalStateException("Invalid Request!");
+        }
         accountService.deleteAccount(customerId, accountType);
     }
 }
